% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptation_tice.R
\name{adaptation_tice}
\alias{adaptation_tice}
\title{Title}
\usage{
adaptation_tice(data, estimates, model, auc_cible)
}
\arguments{
\item{data}{a dataframe, NM-TRAN like}

\item{estimates}{a list, default output of the mapbay_estimation}

\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}

\item{auc_cible}{a numeric input}
}
\value{
a list with results of the adaptation according to Moeung et al, CCR, 2017
}
\description{
Title
}
