% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgmod.R
\name{add_covariates.mrgmod}
\alias{add_covariates.mrgmod}
\title{Add covariates columns to data}
\usage{
\method{add_covariates}{mrgmod}(x, covariates = list(), ...)
}
\arguments{
\item{x}{model object}

\item{covariates}{a list of named covariates, with a single value or exact number of lines than data}

\item{...}{not used}
}
\value{
model object with dataset
}
\description{
Add covariates columns to data
}
