% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_dataset.R
\name{add_random_covariates}
\alias{add_random_covariates}
\title{Generate and add random values of covariates to a dataset (used in validation_dataset)}
\usage{
add_random_covariates(data, model, covariates, random = F)
}
\arguments{
\item{data}{a dataframe, NM-TRAN like}

\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}

\item{covariates}{list of the values of covariates to test (i.e. list(WT = c(60,80)))}

\item{random}{logical. If T, will add an other set of patients with changing covariates (randomly)}
}
\value{
a dataframe, NM-TRAN like
}
\description{
Generate and add random values of covariates to a dataset (used in validation_dataset)
}
