% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_misc_functions.R
\name{add_rate_column}
\alias{add_rate_column}
\title{Title}
\usage{
add_rate_column(data, model, rate = -2)
}
\arguments{
\item{data}{a dataframe, NM-TRAN like}

\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}

\item{rate}{a numeric input, rate of the perfusion. Default value = -2 : duration of infusion expected in the mrgsolve model}
}
\value{
a dataframe, NM-TRAN like
}
\description{
Title
}
