% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_misc_functions.R
\name{addl_to_linebyline}
\alias{addl_to_linebyline}
\title{Title}
\usage{
addl_to_linebyline(data)
}
\arguments{
\item{data}{a dataframe, NM-TRAN like}
}
\value{
a tibble with administration
}
\description{
Title
}
