% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{adm_0_cmt}
\alias{adm_0_cmt}
\title{Get zero-order infusion compartment from mrgsolve model}
\usage{
adm_0_cmt(x)
}
\arguments{
\item{x}{model object}
}
\value{
vector of integer
}
\description{
Get zero-order infusion compartment from mrgsolve model
}
\examples{
model <- mrgsolve::mread("ex_mbr1", mbrlib())
adm_0_cmt(model)
}
