% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgmod.R
\name{adm_lines.mrgmod}
\alias{adm_lines.mrgmod}
\title{Generate Administration lines for a dataset}
\usage{
\method{adm_lines}{mrgmod}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed to mrgsolve::ev}
}
\value{
model object with dataset
}
\description{
Generate Administration lines for a dataset
}
