% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_misc_functions.R
\name{adm_lines}
\alias{adm_lines}
\title{Title}
\usage{
adm_lines(model, time = 0, addl = 0, ii = 0, amt)
}
\arguments{
\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}

\item{time}{a numeric value}

\item{addl}{a numeric value}

\item{ii}{a numeric value}

\item{amt}{a numeric value}
}
\value{
a tibble with administration
}
\description{
Title
}
