% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{as.data.frame.mbrests}
\alias{as.data.frame.mbrests}
\title{Return the mapbay_tab as a data.frame}
\usage{
\method{as.data.frame}{mbrests}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \code{mbrests} object.}

\item{row.names, optional, ...}{passed to as.data.frame}
}
\description{
Return the mapbay_tab as a data.frame
}
