% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbrests.R
\name{augment}
\alias{augment}
\title{Compute full PK profile prediction from mapbayr estimates.}
\usage{
augment(x, ...)
}
\arguments{
\item{x}{object to augment}

\item{...}{additional arguments}
}
\description{
Compute full PK profile prediction from mapbayr estimates.
}
