% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_misc_functions.R
\name{build_input_dataset}
\alias{build_input_dataset}
\title{Title}
\usage{
build_input_dataset(observation_lines, administration_lines)
}
\arguments{
\item{observation_lines}{a tibble with observations}

\item{administration_lines}{a tibble with administration}
}
\value{
a dataframe, NM-TRAN like
}
\description{
Title
}
