% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_simulation_dataset.R
\name{build_simulation_dataset}
\alias{build_simulation_dataset}
\title{Title}
\usage{
build_simulation_dataset(data, sim_amt, sim_ii)
}
\arguments{
\item{data}{a dataframe, NM-TRAN like}

\item{sim_amt}{numeric input}

\item{sim_ii}{numeric input}
}
\value{
a dataframe, NM-TRAN like
}
\description{
Title
}
