% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbayest_check_and_preprocess.R
\name{check_mapbayr_model}
\alias{check_mapbayr_model}
\title{Check if model is valid for mapbayr}
\usage{
check_mapbayr_model(x, check_compile = TRUE)
}
\arguments{
\item{x}{model file}

\item{check_compile}{check if model is compiled (used internally)}
}
\value{
TRUE value if check is passed, a vector of character with errors otherwise.
}
\description{
Check if model is valid for mapbayr
}
\examples{
library(mapbayr)
library(mrgsolve)
check_mapbayr_model(house())
}
