% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbayest_ofv_computation.R
\name{compute_ofv}
\alias{compute_ofv}
\alias{do_compute_ofv}
\title{Compute the objective function value}
\usage{
compute_ofv(
  eta,
  qmod,
  sigma,
  omega_inv,
  ruvdef,
  log_transformation,
  lambda = 1,
  idvaliddata,
  idDV,
  idcmt,
  idblq = NULL,
  idlloq = NULL,
  ...
)

do_compute_ofv(eta, argofv, ...)
}
\arguments{
\item{eta}{a named vector/list of parameters}

\item{qmod, sigma, log_transformation, omega_inv, ruvdef, lambda}{generated by \code{\link{preprocess.ofv.fix}}}

\item{idvaliddata, idDV, idcmt}{generated by \code{\link{preprocess.ofv.id}}}

\item{idblq, idlloq}{optionally generated by \code{\link{preprocess.ofv.id}}}

\item{...}{for compatibility (not used)}

\item{argofv}{above mentioned arguments as a list}
}
\value{
a single numeric value (the objective function value)
}
\description{
Compute the objective function value
}
\details{
This function is called iteratively by the optimization function. Arguments should not be passed directly, but generated by the pre-processing functions (see \code{\link{preprocess.ofv}}).
}
