% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ofv_m2.R
\name{compute_ofv_m}
\alias{compute_ofv_m}
\title{Title}
\usage{
compute_ofv_m(
  eta,
  data,
  mrgsolve_model,
  sigma,
  log.transformation,
  DVobs,
  omega.inv,
  obs_cmt
)
}
\arguments{
\item{eta}{a vector, initial eta, as a named vector (ETA1, ETA2...)}

\item{data}{a tibble, dataset to fit; formatted like NM-TRAN}

\item{mrgsolve_model}{a compiled mrgsolve model}

\item{sigma}{a matrix, population values of sigma. 1 proportional error, 2 additive error}

\item{log.transformation}{logical operator. Useful for proportional error models (i.e. log transformed additive)}

\item{DVobs}{vector of observation (provided by mapbay_estimation to avoid its calculation at each iteration)}

\item{omega.inv}{inverse of omega matix (provided by mapbay_estimation to avoid its calculation at each iteration)}

\item{obs_cmt}{vector of obs cmt}
}
\value{
a single value (the objective function value)
}
\description{
Title
}
