% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prediction.R
\name{compute_prediction}
\alias{compute_prediction}
\title{Title}
\usage{
compute_prediction(data, estimates, model, time_target)
}
\arguments{
\item{data}{a dataframe, NM-TRAN like}

\item{estimates}{a list, default output of the mapbay_estimation}

\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}

\item{time_target}{a numeric, time needed for reestimation of concentration}
}
\value{
a dataframe, a mrgsolve 'df' output
}
\description{
Title
}
