% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_simulation.R
\name{compute_simulation}
\alias{compute_simulation}
\title{Title}
\usage{
compute_simulation(estimates, model, covariates, sim_amt, sim_ii)
}
\arguments{
\item{estimates}{a list, default output of the mapbay_estimation}

\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}

\item{covariates}{a list of named covariates, with a single value or exact number of lines than data}

\item{sim_amt}{numeric input}

\item{sim_ii}{numeric input}
}
\value{
a dataframe, tibble with 5th 50th and 95th percentile of simulation
}
\description{
Title
}
