% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ofv.R
\name{derivatives}
\alias{derivatives}
\title{Title}
\usage{
derivatives(v_DV, v_cmt, cmts)
}
\arguments{
\item{v_DV}{vector of concentrations to derivate (typically output$DV)}

\item{v_cmt}{vector of compartment associated to the concentrations (typically output$cmt)}

\item{cmts}{numbers of compartment compartment associated with an observations, to derive concentration (typically obs_cmt)}
}
\value{
a matrix
}
\description{
Title
}
