% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{est001}
\alias{est001}
\title{Estimation object}
\format{
An object of class \code{mapbayests} of length 9.
}
\usage{
est001
}
\description{
An example of \code{mapbayests} object, corresponding to the parameter estimation of the 8 subjects from model 1. Note that the model object within is not associated to a shared object, which make some features unavailable. This object can be re-generated by executing \code{est001 <- mapbayest(exmodel(ID = 1:8))}
}
\seealso{
\code{\link{mapbayest}}
}
\keyword{datasets}
