% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{filter.mrgmod}
\alias{filter.mrgmod}
\title{Filter a dataset within a mrgmod}
\usage{
\method{filter}{mrgmod}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{a mrgmod}

\item{..., .preserve}{additional arguments for \code{dplyr::filter()}}
}
\value{
a mrgmod
}
\description{
Filter a dataset within a mrgmod
}
\examples{
library(magrittr)
mod <- mrgsolve::mcode("mod", "$CMT FOO", compile = FALSE)
mod \%>\%
  adm_rows(amt = c(100, 200, 300), cmt = 1) \%>\%
  filter(amt != 200) \%>\%
  get_data()
}
