% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_cov.mapbayests}
\alias{get_cov.mapbayests}
\title{Return covariance matrix from a mapbayests}
\usage{
\method{get_cov}{mapbayests}(x, ...)
}
\arguments{
\item{x}{mapbayests object}

\item{...}{not used}
}
\value{
a tibble
}
\description{
Return covariance matrix from a mapbayests
}
