% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_data.mapbayests}
\alias{get_data.mapbayests}
\title{Return data from a mapbayests}
\usage{
\method{get_data}{mapbayests}(x, ..., output = "df")
}
\arguments{
\item{x}{mapbayests object}

\item{...}{not used}

\item{output}{either a single data.frame ("df", the default) or a list ("list") of individual data sets}
}
\value{
a tibble
}
\description{
Return data from a mapbayests
}
