% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_data.mrgmod}
\alias{get_data.mrgmod}
\title{Return data from a mrgmod}
\usage{
\method{get_data}{mrgmod}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{not used}
}
\value{
a tibble
}
\description{
Return data from a mrgmod
}
