% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_eta.mapbayests}
\alias{get_eta.mapbayests}
\title{Return eta from a mapbayests}
\usage{
\method{get_eta}{mapbayests}(x, ..., output = NULL)
}
\arguments{
\item{x}{mapbayests object}

\item{...}{not used}

\item{output}{either a list ("list"), a data.frame ("df") or a vector of numeric ("num"). Default to "num" if only one ID.}
}
\value{
a tibble
}
\description{
Return eta from a mapbayests
}
