% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_param.mapbayests}
\alias{get_param.mapbayests}
\title{Return a posteriori param from a mapbayests}
\usage{
\method{get_param}{mapbayests}(x, ...)
}
\arguments{
\item{x}{mapbayests object}

\item{...}{passed along}
}
\value{
a tibble
}
\description{
Return a posteriori param from a mapbayests
}
