% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_phi.mapbayests}
\alias{get_phi.mapbayests}
\title{Return "NONMEM phi"-like file from a mapbayests}
\usage{
\method{get_phi}{mapbayests}(x, ...)
}
\arguments{
\item{x}{mapbayests object}

\item{...}{not used}
}
\value{
a tibble
}
\description{
Return "NONMEM phi"-like file from a mapbayests
}
