% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_x}
\alias{get_x}
\alias{get_data}
\alias{get_eta}
\title{Get content from object}
\usage{
get_data(x, ...)

get_eta(x, ...)
}
\arguments{
\item{x}{object to get content from}

\item{...}{passed along}
}
\description{
Helpful functions to get content from a \code{mrgmod} object (i.e. data) or from a \code{mapbayests} object (data, eta, param).
}
