% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbayests.R
\name{hist.mapbayests}
\alias{hist.mapbayests}
\title{Plot posterior distribution of bayesian estimates}
\usage{
\method{hist}{mapbayests}(x, ...)
}
\arguments{
\item{x}{A \code{mapbayests} object.}

\item{...}{additional arguments (not used)}
}
\description{
Plot posterior distribution of bayesian estimates
}
