% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbrests.R
\name{hist.mbrests}
\alias{hist.mbrests}
\title{Plot posterior distribution of bayesian estimates}
\usage{
\method{hist}{mbrests}(x, ...)
}
\arguments{
\item{x}{A \code{mbrests} object.}

\item{...}{additional arguments (not used)}
}
\description{
Plot posterior distribution of bayesian estimates
}
