% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_misc_functions.R
\name{hour_to_time}
\alias{hour_to_time}
\title{Title}
\usage{
hour_to_time(h0, ..., date = "1970/01/01")
}
\arguments{
\item{h0}{character input, initial hour}

\item{...}{character input, hours to calculate difference from h0}

\item{date}{character input, optionnal date}
}
\value{
vector of numeric
}
\description{
Title
}
