% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_misc_functions.R
\name{lag_covariates}
\alias{lag_covariates}
\title{Title}
\usage{
lag_covariates(data, covariates)
}
\arguments{
\item{data}{a dataframe, NM-TRAN like}

\item{covariates}{a vector of characters, names of the covariates column to "lag"}
}
\value{
a dataframe, NM-TRAN like
}
\description{
Title
}
