% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mapbay_model.R
\name{load_mapbay_model}
\alias{load_mapbay_model}
\title{Title}
\usage{
load_mapbay_model(model, path = NULL)
}
\arguments{
\item{model}{character name of the model in the mapbay model library of the package}

\item{path}{exact path to the model. Default = NULL (read in the library of mapbay models)}
}
\value{
a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model
}
\description{
Title
}
