% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbay_estimation2.R
\name{mapbay_estimation2}
\alias{mapbay_estimation2}
\title{Title}
\usage{
mapbay_estimation2(
  data,
  model,
  output = "default",
  force_initial_eta = NULL,
  sigdig = 3,
  force_rtol = -8,
  control_optim = list(maxit = 50000, pgtol = 10^-sigdig, trace = 6)
)
}
\arguments{
\item{data}{a dataframe, dataset (NM-TRAN format) of one individual to fit}

\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}

\item{output}{a string. If output = "df" a dataframe is returned and not a list of component. Useful for mapping multiple individual datasets}

\item{force_initial_eta}{a numeric vector of starting estimates (exact length of eta to estimate )}

\item{sigdig}{a positive integer (cf focei nlmixr for significance)}

\item{force_rtol}{change rtol of the mrgsolve model}

\item{control_optim}{list of argument to be passed to optim function}
}
\value{
default: a list with initial eta, newuoa outputs, final eta and a NM-like dataset with predictions
}
\description{
Title
}
