% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbay_estimation2.R
\name{mapbay_estimation3}
\alias{mapbay_estimation3}
\title{Title}
\usage{
mapbay_estimation3(
  data,
  model,
  output = "default",
  force_initial_eta = NULL,
  force_rtol = -8,
  control_nloptr = list(algorithm = "NLOPT_LN_NEWUOA", maxeval = 50000, print_level =
    1)
)
}
\arguments{
\item{data}{a dataframe, dataset (NM-TRAN format) of one individual to fit}

\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}

\item{output}{a string. If output = "df" a dataframe is returned and not a list of component. Useful for mapping multiple individual datasets}

\item{force_initial_eta}{a numeric vector of starting estimates (exact length of eta to estimate )}

\item{force_rtol}{change rtol of the mrgsolve model}

\item{control_nloptr}{list of argument to be passed to nloptr function}
}
\value{
default: a list with initial eta, newuoa outputs, final eta and a NM-like dataset with predictions
}
\description{
Title
}
