% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbay_library.R
\name{mapbay_library}
\alias{mapbay_library}
\title{Internal library of mapbay models}
\usage{
mapbay_library(
  path = system.file("mrg_models", package = "mapbayr"),
  app_list = F
)
}
\arguments{
\item{path}{path to the internal library}

\item{app_list}{if TRUE, return a list used in ui.R of tki app}
}
\value{
a data.frame
}
\description{
Internal library of mapbay models
}
