% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{mbr_conc_scaling}
\alias{mbr_conc_scaling}
\title{Get scaling factor from user to model}
\usage{
mbr_conc_scaling(x)
}
\arguments{
\item{x}{model object}
}
\value{
an integer
}
\description{
Get scaling factor from user to model
}
\examples{
#Dose expected in mg, and Volume of distribution in L.
#Possible values for concentration units: mg/L, ng/mL, pg/mL
#Microgramme not supported for compatibility and encoding reasons (greek letter mu)
model <- mrgsolve::mread("ex_mbr1", mbrlib())
mbr_conc_scaling(model)
}
