% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{mbr_conc_unit}
\alias{mbr_conc_unit}
\title{Get concentration units from mrgsolve model}
\usage{
mbr_conc_unit(x)
}
\arguments{
\item{x}{model object}
}
\value{
vector of character
}
\description{
Get concentration units from mrgsolve model
}
\examples{
model <- mrgsolve::mread(system.file("mrg_models/example_test_mapbay.cpp", package = "mapbayr"))
mbr_conc_unit(model)
}
