% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{mbr_cov_descr}
\alias{mbr_cov_descr}
\title{Get covariate description from mrgsolve model}
\usage{
mbr_cov_descr(x)
}
\arguments{
\item{x}{model object}
}
\value{
a character string vector
}
\description{
Get covariate description from mrgsolve model
}
\examples{
model <- mrgsolve::mread("ex_mbr1", mbrlib())
mbr_cov_descr(model)
}
