% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{mbr_cov_names}
\alias{mbr_cov_names}
\title{Get covariate names from mrgsolve model}
\usage{
mbr_cov_names(x)
}
\arguments{
\item{x}{model object}
}
\value{
a character string vector
}
\description{
Get covariate names from mrgsolve model
}
\examples{
model <- mrgsolve::mread(system.file("mrg_models/example_test_mapbay.cpp", package = "mapbayr"))
mbr_cov_names(model)
}
