% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{mbr_cov_refvalues}
\alias{mbr_cov_refvalues}
\title{Get covariate reference values from mrgsolve model}
\usage{
mbr_cov_refvalues(x)
}
\arguments{
\item{x}{model object}
}
\value{
a named vector of numerics
}
\description{
Get covariate reference values from mrgsolve model
}
\examples{
model <- mrgsolve::mread("ex_mbr1", mbrlib())
mbr_cov_refvalues(model)
}
