% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{mbr_drug}
\alias{mbr_drug}
\title{Get drug name from mrgsolve model}
\usage{
mbr_drug(x)
}
\arguments{
\item{x}{model object}
}
\value{
name of the drug as a character string
}
\description{
Get drug name from mrgsolve model
}
\examples{
model <- mrgsolve::mread("ex_mbr1", mbrlib())
mbr_drug(model)

}
