% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{mbr_eta_names}
\alias{mbr_eta_names}
\title{Get eta names from mrgsolve model}
\usage{
mbr_eta_names(x)
}
\arguments{
\item{x}{model object}
}
\value{
a character string vector
}
\description{
Get eta names from mrgsolve model
}
\examples{
model <- mrgsolve::mread("ex_mbr1", mbrlib())
mbr_eta_names(model)
}
