% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{mbr_model_file}
\alias{mbr_model_file}
\title{Get model file from mrgsolve model}
\usage{
mbr_model_file(x)
}
\arguments{
\item{x}{model object}
}
\value{
model file as a character string
}
\description{
Get model file from mrgsolve model
}
\examples{
model <- mrgsolve::mread(system.file("mrg_models/example_test_mapbay.cpp", package = "mapbayr"))
mbr_model_name(model)

}
