% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{mbr_model_name}
\alias{mbr_model_name}
\title{Get model name from mrgsolve model}
\usage{
mbr_model_name(x)
}
\arguments{
\item{x}{model object}
}
\value{
model name as a character string
}
\description{
Get model name from mrgsolve model
}
\examples{
model <- mrgsolve::mread("ex_mbr1", mbrlib())
mbr_model_name(model)

}
