% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{mbr_param_names}
\alias{mbr_param_names}
\title{Get parameter names from mrgsolve model}
\usage{
mbr_param_names(x)
}
\arguments{
\item{x}{model object}
}
\value{
a character string vector
}
\description{
Get parameter names from mrgsolve model
}
\examples{
model <- mrgsolve::mread("ex_mbr1", mbrlib())
mbr_param_names(model)
}
