% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{mbr_param_tv}
\alias{mbr_param_tv}
\title{Get parameter typical values from mrgsolve model}
\usage{
mbr_param_tv(x)
}
\arguments{
\item{x}{model object}
}
\value{
a numeric vector
}
\description{
Get parameter typical values from mrgsolve model
}
\examples{
model <- mrgsolve::mread(system.file("mrg_models/example_test_mapbay.cpp", package = "mapbayr"))
mbr_param_tv(model)
}
