% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{mbr_param_units}
\alias{mbr_param_units}
\title{Get parameter units from mrgsolve model}
\usage{
mbr_param_units(x)
}
\arguments{
\item{x}{model object}
}
\value{
a character string vector
}
\description{
Get parameter units from mrgsolve model
}
\examples{
model <- mrgsolve::mread(system.file("mrg_models/example_test_mapbay.cpp", package = "mapbayr"))
mbr_param_units(model)
}
