% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploit_estimates.R
\name{mbrdist}
\alias{mbrdist}
\title{Plot posterior distribution of bayesian estimates}
\usage{
mbrdist(est)
}
\arguments{
\item{est}{a list, default output of mbrest}
}
\value{
a ggplot-object
}
\description{
Plot posterior distribution of bayesian estimates
}
