% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbay_estimation.R
\name{mbrest}
\alias{mbrest}
\title{Perform mapbayesian estimation from a mrgsolve model and a NM_tran dataset}
\usage{
mbrest(x, data = NULL, output = NULL, force_initial_eta = NULL)
}
\arguments{
\item{x}{model file}

\item{data}{NM tran data to optimize}

\item{output}{return a mapbay_tab only}

\item{force_initial_eta}{a numeric vector of starting estimates (exact length of eta to estimate )}
}
\value{
default: a list with data, model, initial eta, newuoa outputs, final eta, and mapbay_tab
}
\description{
Perform mapbayesian estimation from a mrgsolve model and a NM_tran dataset
}
