% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{mbrlib}
\alias{mbrlib}
\title{Internal "mapbayr" model examples}
\usage{
mbrlib()
}
\description{
Internal "mapbayr" model examples
}
\examples{
\dontrun{
mod <- mrgsolve::mread("ex_mbr1", mbrlib())
mod <- mrgsolve::mread("ex_mbr2", mbrlib())
mod <- mrgsolve::mread("ex_mbr3", mbrlib())
}
}
