% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploit_estimates.R
\name{mbrplot}
\alias{mbrplot}
\title{Plot predictions from mapbayr estimates}
\usage{
mbrplot(est)
}
\arguments{
\item{est}{a list, default output of mbrest augmented with mbrpred}
}
\value{
a ggplot-object
}
\description{
Plot predictions from mapbayr estimates
}
