% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{obs_cmt}
\alias{obs_cmt}
\title{Get observation compartment numbers from mrgsolve model}
\usage{
obs_cmt(x)
}
\arguments{
\item{x}{model object}
}
\value{
vector of integer
}
\description{
Get observation compartment numbers from mrgsolve model
}
\examples{
model <- mrgsolve::mread("ex_mbr1", mbrlib())
obs_cmt(model)
}
