% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgmod.R
\name{obs_lines.mrgmod}
\alias{obs_lines.mrgmod}
\title{Add observations lines to data}
\usage{
\method{obs_lines}{mrgmod}(x, time, DV, mdv = 0, cmt = NULL, DVmet = NULL, ...)
}
\arguments{
\item{x}{model object}

\item{time}{vector of time}

\item{DV}{vector of values to fit}

\item{mdv}{optional should the DV be ignored (1) or not (0)}

\item{cmt}{an integer. Optional, cmt where observation is measured}

\item{DVmet}{optional : metabolite data to fit}

\item{...}{not used}
}
\value{
model object with dataset
}
\description{
Add observations lines to data
}
