% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_misc_functions.R
\name{obs_lines}
\alias{obs_lines}
\title{Title}
\usage{
obs_lines(model, time, DV, mdv = 0)
}
\arguments{
\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}

\item{time}{a numeric vector of time after the last dose}

\item{DV}{a numeric vector of observation}

\item{mdv}{a numeric vector of mdv to ignore observation}
}
\value{
a tibble with observations
}
\description{
Title
}
