% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbrest_postprocess.R
\name{output_mbr}
\alias{output_mbr}
\title{Build the output of mbrest function}
\usage{
output_mbr(x, data, arg.optim, arg.ofv, opt.value, post, output)
}
\arguments{
\item{x}{the model object}

\item{data}{NMTRAN-like data set}

\item{arg.optim}{argument passed to optimizer}

\item{arg.ofv}{argument passed to optimizer}

\item{opt.value}{value returned by optimizer}

\item{post}{output of the post.process function}

\item{output}{if \code{NULL} (the default) a mbrests object is returned; if \code{df} a \emph{mapbay_tab} dataframe is returned}
}
\value{
a mbrests model object
}
\description{
Build the output of mbrest function
}
