% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbayests.R
\name{plot.mapbayests}
\alias{plot.mapbayests}
\title{Plot predictions from mapbayests object}
\usage{
\method{plot}{mapbayests}(x, ...)
}
\arguments{
\item{x}{A \code{mapbayests} object.}

\item{...}{additional arguments (not used)}
}
\description{
Plot predictions from mapbayests object
}
