% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbrests.R
\name{plot.mbrests}
\alias{plot.mbrests}
\title{Plot predictions from mbrests object}
\usage{
\method{plot}{mbrests}(x, ...)
}
\arguments{
\item{x}{A \code{mbrests} object.}

\item{...}{additional arguments (not used)}
}
\description{
Plot predictions from mbrests object
}
