% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log_normal_distribution.R
\name{plot_log_normal_distribution}
\alias{plot_log_normal_distribution}
\title{Title}
\usage{
plot_log_normal_distribution(PAR, TV, OM, EBE, UNIT)
}
\arguments{
\item{PAR}{character, name of the parameter}

\item{TV}{numeric, typical value of the paremeter}

\item{OM}{numeric, diagonal value of omega}

\item{EBE}{numeric, empirical bayesian estimate value}

\item{UNIT}{character, unit of the parameter}
}
\value{
a ggplot-object
}
\description{
Title
}
