% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pred_mapbay.R
\name{plot_pred_mapbay}
\alias{plot_pred_mapbay}
\title{Title}
\usage{
plot_pred_mapbay(
  pred_data,
  input_data,
  model,
  from = 0,
  to = max(pred_data$time),
  hline = NULL
)
}
\arguments{
\item{pred_data}{a dataframe, NM-TRAN like}

\item{input_data}{a dataframe, NM-TRAN like}

\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}

\item{from}{a numeric input}

\item{to}{a numeric input}

\item{hline}{a numeric input, can be provided by interpretation_tki$concentration_target}
}
\value{
a ggplot-object
}
\description{
Title
}
