% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_tki.R
\name{plot_sim_tki}
\alias{plot_sim_tki}
\title{Title}
\usage{
plot_sim_tki(sim_data, model, interpretation_tki, zoom)
}
\arguments{
\item{sim_data}{a dataframe, tibble with 5th 50th and 95th percentile of simulation}

\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}

\item{interpretation_tki}{a list, characteristic of the drug, with therapeutic target}

\item{zoom}{zoom input (shiny app button)}
}
\value{
a ggplot-object
}
\description{
Title
}
