% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbrest_postprocess.R
\name{postprocess.optim}
\alias{postprocess.optim}
\title{Post-process: derive predictions from optimization}
\usage{
postprocess.optim(x, data, opt.value, arg.optim, arg.ofv)
}
\arguments{
\item{x}{the model object}

\item{data}{NMTRAN-like data set}

\item{opt.value}{value returned by optimizer}

\item{arg.optim, arg.ofv}{argument passed to optimizer}
}
\value{
a list of post processing values
}
\description{
Post-process: derive predictions from optimization
}
