% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbrest_postprocess.R
\name{postprocess.output}
\alias{postprocess.output}
\title{Post-process: Build the output (mbrests model object)}
\usage{
postprocess.output(x, data, arg.optim, arg.ofv, opt.value, post, output)
}
\arguments{
\item{x}{the model object}

\item{data}{NMTRAN-like data set}

\item{arg.optim}{argument passed to optimizer}

\item{arg.ofv}{argument passed to optimizer}

\item{opt.value}{value returned by optimizer}

\item{post}{output of the post.process function}

\item{output}{if \code{NULL} (the default) a mbrests object is returned; if \code{df} a \emph{mapbay_tab} dataframe is returned}
}
\value{
a mbrests model object
}
\description{
Post-process: Build the output (mbrests model object)
}
