% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_post_process.R
\name{postprocess}
\alias{postprocess}
\title{Post process results from optimization}
\usage{
postprocess(x, data, opt.value, arg.optim, arg.ofv)
}
\arguments{
\item{x}{the model object}

\item{data}{NMTRAN-like data set}

\item{opt.value}{value returned by optimizer}

\item{arg.optim, arg.ofv}{argument passed to optimizer}
}
\value{
a list of post processing values
}
\description{
Post process results from optimization
}
