% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_post_process.R
\name{preprocess.data}
\alias{preprocess.data}
\title{Preprocess: data into a list of individual data}
\usage{
preprocess.data(data)
}
\arguments{
\item{data}{a NM-TRAN data set of one or multiple individuals.}
}
\value{
a named list of n data set (n individuals)
}
\description{
Preprocess: data into a list of individual data
}
