% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbay_estimation.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess model and data for optimization}
\usage{
preprocess(model, data, force_initial_eta = NULL)
}
\arguments{
\item{model}{a compiled mrgsolve_model}

\item{data}{a dataframe, dataset (NM-TRAN format) of one individual to fit}

\item{force_initial_eta}{a numeric vector of starting estimates (exact length of eta to estimate )}
}
\value{
a data.frame ready for mapbay_estimation
}
\description{
Preprocess model and data for optimization
}
