% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.mbrests}
\alias{print.mbrests}
\title{Print a mbrests object}
\usage{
\method{print}{mbrests}(x, ...)
}
\arguments{
\item{x}{A \code{mbrests} object.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print a mbrests object
}
