% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbay_library.R
\name{print_code}
\alias{print_code}
\title{Print code of .cpp file}
\usage{
print_code(model)
}
\arguments{
\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}
}
\value{
print code
}
\description{
Print code of .cpp file
}
