% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_data_tice.R
\name{print_data_tice}
\alias{print_data_tice}
\title{Title}
\usage{
print_data_tice(input_data, estimates)
}
\arguments{
\item{input_data}{a dataframe, NM-TRAN like}

\item{estimates}{a list, default output of the mapbay_estimation}
}
\value{
a dataframe, NM-TRAN like
}
\description{
Title
}
