% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_data_tki.R
\name{print_data_tki}
\alias{print_data_tki}
\title{Title}
\usage{
print_data_tki(data)
}
\arguments{
\item{data}{a dataframe, NM-TRAN like}
}
\value{
a dataframe, NM-TRAN like (except colnames)
}
\description{
Title
}
