% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_misc_functions.R
\name{rate_duration}
\alias{rate_duration}
\title{Title}
\usage{
rate_duration(amt, start, end)
}
\arguments{
\item{amt}{a numeric input}

\item{start}{a numeric input}

\item{end}{a numeric input}
}
\value{
a list of numerics
}
\description{
Title
}
