% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgmod.R
\name{see_data.mrgmod}
\alias{see_data.mrgmod}
\title{Return data}
\usage{
\method{see_data}{mrgmod}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{not used}
}
\value{
a tibble
}
\description{
Return data
}
