% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{see_data}
\alias{see_data}
\title{Print tibble data to the console}
\usage{
see_data(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed to print}
}
\value{
called for its side effect
}
\description{
Print tibble data to the console
}
