% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_estimates.R
\name{use_estimates}
\alias{use_estimates}
\title{Use parameter estimates}
\usage{
use_estimates(
  x,
  use_eta = TRUE,
  use_covariates = TRUE,
  .etasrc = "idata.all",
  .zero_re = "both",
  verbose = TRUE
)
}
\arguments{
\item{x}{A \code{mapbayests} object.}

\item{use_eta}{a logical. Populate the \code{idataset} with the estimates (ETA).}

\item{use_covariates}{a logical. Populate the \code{idataset} with the covariate values available in the data.}

\item{.etasrc}{a character. Value used to populate \code{mrgsim(etasrc)}.}

\item{.zero_re}{a character. Set all elements of the OMEGA or SIGMA matrix to zero. Default is "both", alternatively "sigma", "omega" and "none".}

\item{verbose}{a logical. Display information to the console.}
}
\value{
An mrgsolve model object (of class "mrgmod").
}
\description{
Use parameter estimates
}
\details{
This function takes the results of an estimation (i.e. a \code{mapbayests} object) and return a modified \code{mrgmod} in order to perform \emph{a posteriori} simulations. Modifications are:
\itemize{
\item An individual data set (\code{idata_set}), populated with the estimated ETA parameters and the covariate values provided in the original data set.
\item The argument \code{etasrc} set to "idata.all" (or different, depending on the value set to \code{.etasrc}).
\item OMEGA and SIGMA matrices set to zero (or different, depending on the value of \code{.zero_re}).
It does not handle time-varying covariates: only the first value will be used as the individual value.
}
}
\examples{
library(magrittr)
library(mrgsolve)
est <- mapbayest(exmodel(ID = 1:8))
est \%>\%
  use_estimates() \%>\%
  ev(amt = 50000) \%>\%
  mrgsim() \%>\%
  plot()
}
