% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbayests.R
\name{use_posterior}
\alias{use_posterior}
\title{Use posterior param and covariates}
\usage{
use_posterior(x, .zero_re = c("both", "omega", "sigma"))
}
\arguments{
\item{x}{A \code{mapbayests} object.}

\item{.zero_re}{Default is "both", meaning all matrices are zeroed. Pass "omega" to zero between-subject variability, and keep simulating residual error.}
}
\value{
a mrgmod
}
\description{
Use posterior param and covariates
}
\details{
Updates the param values of the model object with the estimated etas, and the covariates of the individual. Returns an updated mrgmod, so that the user can derive simulations from it. Works only with one individual. Does not handle time-varying covariates.
}
