% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_dataset.R
\name{validation_dataset}
\alias{validation_dataset}
\title{Create a validation dataset (comparison vs NONMEM)}
\usage{
validation_dataset(
  model,
  min_n_day_ttt,
  max_n_day_ttt,
  by_n_day_ttt,
  dosing_ii,
  amts,
  change_dose = T,
  n_random_samples_1,
  n_random_samples_2,
  covariates = NULL,
  random = F,
  nrep = 1
)
}
\arguments{
\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}

\item{min_n_day_ttt}{numeric input}

\item{max_n_day_ttt}{numeric input}

\item{by_n_day_ttt}{numeric input}

\item{dosing_ii}{numeric input}

\item{amts}{numeric input (vector)}

\item{change_dose}{a logical value}

\item{n_random_samples_1}{numeric input}

\item{n_random_samples_2}{numeric input}

\item{covariates}{list of the values of covariates to test (i.e. list(WT = c(60,80)))}

\item{random}{logical. If T, will add an other set of patients with changing covariates (randomly)}

\item{nrep}{numeric input (n replicates)}
}
\value{
a dataframe, NM-TRAN like
}
\description{
Create a validation dataset (comparison vs NONMEM)
}
