% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tki.R
\name{write_comment_biology_tki}
\alias{write_comment_biology_tki}
\title{Title}
\usage{
write_comment_biology_tki(pred_data, model, interpretation_tki, time_target)
}
\arguments{
\item{pred_data}{a dataframe, NM-TRAN like}

\item{model}{a list, characteristics of the model, including a compiled mrgsolve_model}

\item{interpretation_tki}{a list, characteristic of the drug, with therapeutic target}

\item{time_target}{numeric input}
}
\value{
a character string
}
\description{
Title
}
