% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tki.R
\name{write_comment_fit_tki}
\alias{write_comment_fit_tki}
\title{Title}
\usage{
write_comment_fit_tki(pred_data, estimates, model, time_last_dose, time_target)
}
\arguments{
\item{pred_data}{a dataframe, NM-TRAN like}

\item{estimates}{a list, default output of the mapbay_estimation}

\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}

\item{time_last_dose}{numeric input}

\item{time_target}{numeric input}
}
\value{
a character string
}
\description{
Title
}
