% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tice.R
\name{write_comment_tice}
\alias{write_comment_tice}
\title{Title}
\usage{
write_comment_tice(adapation, n_cycle, auc_cible, date_adm)
}
\arguments{
\item{adapation}{a list, output of adaptation_tice}

\item{n_cycle}{a numeric input}

\item{auc_cible}{a numeric input}

\item{date_adm}{a date input}
}
\value{
a character string
}
\description{
Title
}
