% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tice.R
\name{write_result_tice}
\alias{write_result_tice}
\title{Title}
\usage{
write_result_tice(adapation, n_cycle)
}
\arguments{
\item{adapation}{a list, output of adaptation_tice}

\item{n_cycle}{a numeric input}
}
\value{
a character string
}
\description{
Title
}
