% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgmod.R
\name{add_covariates.mrgmod}
\alias{add_covariates.mrgmod}
\title{Add covariates to the dataset, as well as TOLA and AOLA}
\usage{
\method{add_covariates}{mrgmod}(model, covariates = list())
}
\arguments{
\item{model}{model object}

\item{covariates}{a list of named covariates, with a single value or exact number of lines than data}
}
\value{
model object with dataset
}
\description{
Add covariates to the dataset, as well as TOLA and AOLA
}
