% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_misc_functions.R
\name{add_covariates}
\alias{add_covariates}
\title{Title}
\usage{
add_covariates(data, model, covariates = list())
}
\arguments{
\item{data}{a dataframe, NM-TRAN like}

\item{model}{a mapbay_model list, with the characteristics of the model, including a compiled mrgsolve_model}

\item{covariates}{a list of named covariates, with a single value or exact number of lines than data}
}
\value{
a dataframe, NM-TRAN like
}
\description{
Title
}
