% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{adm_lines}
\alias{adm_lines}
\title{Generate Administration lines for a dataset}
\usage{
adm_lines(
  model,
  time = 0,
  addl = 0,
  ii = 0,
  amt = 0,
  rate = 0,
  realize_addl = F,
  output = NULL
)
}
\arguments{
\item{model}{model object}

\item{time}{a numeric value}

\item{addl}{a numeric value}

\item{ii}{a numeric value}

\item{amt}{a numeric value}

\item{rate}{a numeric value (for IV only. Automatically filled with -2 if zero order set in model code)}

\item{realize_addl}{a logical (see mrgsolve::realize_addl)}

\item{output}{defaut : a mrgsolve model with a data_set args. if "df" return a data frame}
}
\value{
model object with dataset
}
\description{
Generate Administration lines for a dataset
}
