% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{adm_lines}
\alias{adm_lines}
\title{Generate Administration lines for a dataset}
\usage{
adm_lines(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed to mrgsolve::ev}
}
\value{
model object with dataset
}
\description{
Generate Administration lines for a dataset
}
