% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbayests.R
\name{as.data.frame.mapbayests}
\alias{as.data.frame.mapbayests}
\title{Return the mapbay_tab as a data.frame}
\usage{
\method{as.data.frame}{mapbayests}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \code{mapbayests} object.}

\item{row.names, optional, ...}{passed to as.data.frame}
}
\description{
Return the mapbay_tab as a data.frame
}
