% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbayests.R
\name{augment.mapbayests}
\alias{augment.mapbayests}
\title{Compute full PK profile prediction from mapbayr estimates.}
\usage{
\method{augment}{mapbayests}(x, data = NULL, end = NULL, ...)
}
\arguments{
\item{x}{A \code{mapbayests} object.}

\item{data}{dataset to pass to mrgsolve for simulation (default is dataset used for estimation)}

\item{end}{end of infusion time (passed to mrgsim)}

\item{...}{additional argument to pass to mrgsim}
}
\value{
a \code{mapbayests} object, augmented of an \code{aug_tab}
}
\description{
Compute full PK profile prediction from mapbayr estimates.
}
