% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbrest_ofv_computation.R
\name{compute_ofv}
\alias{compute_ofv}
\title{Compute the objective function values. Called iteratively during minimization}
\usage{
compute_ofv(
  eta,
  mrgsolve_model,
  data,
  sigma,
  log.transformation,
  DVobs,
  omega.inv,
  obs_cmt
)
}
\arguments{
\item{eta}{a vector, initial eta, as a named vector (ETA1, ETA2...)}

\item{mrgsolve_model}{a compiled mrgsolve model}

\item{data}{data set for administration and cmt/mdv data}

\item{sigma}{a matrix, population values of sigma. 1 proportional error, 2 additive error}

\item{log.transformation}{logical operator. Useful for proportional error models (i.e. log transformed additive)}

\item{DVobs}{vector of observation (provided by mapbay_estimation to avoid its calculation at each iteration)}

\item{omega.inv}{inverse of omega matix (provided by mapbay_estimation to avoid its calculation at each iteration)}

\item{obs_cmt}{vector of obs cmt (not used)}
}
\value{
a single value (the objective function value)
}
\description{
Compute the objective function values. Called iteratively during minimization
}
