% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{data_helpers}
\alias{data_helpers}
\alias{adm_lines}
\alias{adm_lines.data.frame}
\alias{adm_lines.missing}
\alias{adm_lines.mrgmod}
\alias{obs_lines}
\alias{obs_lines.data.frame}
\alias{obs_lines.missing}
\alias{obs_lines.mrgmod}
\alias{add_covariates}
\alias{add_covariates.data.frame}
\alias{add_covariates.mrgmod}
\title{Data helpers}
\usage{
adm_lines(x, ...)

\method{adm_lines}{data.frame}(
  x,
  ID = NULL,
  time = NULL,
  evid = 1L,
  cmt,
  amt,
  mdv = 1L,
  addl = NULL,
  ss = NULL,
  ii = NULL,
  rate = NULL,
  ...
)

\method{adm_lines}{missing}(...)

\method{adm_lines}{mrgmod}(x, cmt = adm_cmt(x), rate = NULL, ...)

obs_lines(x, ...)

\method{obs_lines}{data.frame}(x, ID = NULL, time, evid = 0L, cmt, DV = NA_real_, mdv = NULL, ...)

\method{obs_lines}{missing}(...)

\method{obs_lines}{mrgmod}(x, cmt = NULL, DV = NA_real_, DVmet = NULL, ...)

add_covariates(x, ...)

\method{add_covariates}{data.frame}(x, ..., covariates = list(), AOLA = FALSE, TOLA = FALSE)

\method{add_covariates}{mrgmod}(x, ..., covariates = list(), AOLA = NULL, TOLA = NULL)
}
\arguments{
\item{x}{either a data.frame or a 'mrgsolve' model object}

\item{...}{additional columns or arguments for \code{\link[mrgsolve:ev]{mrgsolve::ev()}}}

\item{ID}{subject ID (default is 1)}

\item{time}{event time}

\item{evid}{event identification (default is 1 for administration, 0 for observation)}

\item{cmt}{compartment (no default, except if one \verb{[ADM]} and \verb{[OBS]} were tagged in the \verb{$CMT} block in model code. See \code{details}.)}

\item{amt}{dose amount (for administration records only)}

\item{mdv}{missing dependant value (default is 0 for observation to take into account for parameter estimation, 1 otherwise)}

\item{addl}{additional dose (optional and for administration records only)}

\item{ss}{steady-state (optional and for administration records only. Is this dose the last of an infinity of administration? Yes, 1, or no, 0)}

\item{ii}{interdose interval (optional and for administration records only. Use it with \code{ss} and \code{addl})}

\item{rate}{rate of administration (optional and for administration records only. Set to -2 if you model zero-order infusion. See \code{details}.)}

\item{DV}{dependant value, i.e. observed concentration (for observation records only)}

\item{DVmet}{passed to \code{obs_lines()}}

\item{covariates}{a list of named covariates, with a single value or same number of lines than data}
}
\value{
a \code{mrgmod} object, with a dataset in the \verb{@args$data} slot.
}
\description{
Create or modify a dataset from scratch, from a pre-existing dataset, or from a dataset stored into a 'mrgsolve' model
}
\details{
Helpful functions build the data set. Instead of painfully build a data set and mind how to format the data, you can pass information about :
\itemize{
\item administrations with \code{adm_lines()},
\item observations with \code{obs_lines()}
\item covariates with \code{add_covariates()}.
}

These functions are passed to a \code{mrgmod} object (mrgsolve model), and return a \code{mrgmod} object with a data set inside with the correct formatting (so-called NM-TRAN format), so that mrgsolve or mapbayr functions can be passed along within a pipe-friendly workflow.

These functions are meant to be used for one single patient at a time. Multiple ID is accepted, but the user is asked to check if the output is acceptable.
}
\examples{
library(magrittr)
# First, import a model
mod <- exmodel(add_exdata = FALSE)

mod \%>\%
  adm_lines(amt = 10000, cmt = 1) \%>\%
  obs_lines(time = c(1.5, 4.4, 7.5, 24.6), DV = c(91.2904, 110.826, 79.384, 20.6671), cmt = 2) \%>\%
  # get_data() # for curiosity, you can extract the data set at this step
  mapbayest()

# If `[ADM]` or `[OBS]` are set in `$CMT` in model code, the `cmt =` argument are superfluous.

}
