% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{data_helpers}
\alias{data_helpers}
\title{Data helpers}
\value{
a \code{mrgmod} object, with a dataset in the \verb{@args$data} slot.
}
\description{
Create or modify a dataset from scratch, from a pre-existing dataset, or from a dataset stored into a 'mrgsolve' model
}
\details{
Helpful functions build the data set. Instead of painfully build a data set and mind how to format the data, you can pass information about :
\itemize{
\item administrations with \code{adm_lines()},
\item observations with \code{obs_lines()}
\item covariates with \code{add_covariates()}.
}

These functions are passed to a \code{mrgmod} object (mrgsolve model), and return a \code{mrgmod} object with a data set inside with the correct formatting (so-called NM-TRAN format), so that mrgsolve or mapbayr functions can be passed along within a pipe-friendly workflow.

These functions are meant to be used for one single patient at a time. Multiple ID is accepted, but the user is asked to check if the output is acceptable.
}
\examples{
library(magrittr)
# First, import a model
mod <- exmodel(add_exdata = FALSE)

mod \%>\%
  adm_lines(amt = 10000, cmt = 1) \%>\%
  obs_lines(time = c(1.5, 4.4, 7.5, 24.6), DV = c(91.2904, 110.826, 79.384, 20.6671), cmt = 2) \%>\%
  # get_data() # for curiosity, you can extract the data set at this step
  mapbayest()

# If `[ADM]` or `[OBS]` are set in `$CMT` in model code, the `cmt =` argument are superfluous.

}
