% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{deprecations}
\alias{deprecations}
\alias{mbrest}
\alias{adm_lines}
\alias{obs_lines}
\title{Deprecated functions}
\usage{
mbrest(...)

adm_lines(...)

obs_lines(...)
}
\arguments{
\item{...}{passed to the corresponding function}
}
\description{
Deprecated functions
}
\details{
\itemize{
\item mbrest() is now \code{\link[=mapbayest]{mapbayest()}}
\item adm_lines() is now \code{\link[=adm_rows]{adm_rows()}}
\item obs_lines() is now \code{\link[=obs_rows]{obs_rows()}}
}
}
