% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbayest_ofv_computation.R
\name{derivatives}
\alias{derivatives}
\title{Compute the derivatives}
\usage{
derivatives(v_DV, v_cmt, cmts)
}
\arguments{
\item{v_DV}{vector of concentrations to derive (typically output$DV)}

\item{v_cmt}{vector of compartment associated to the concentrations (typically output$cmt)}

\item{cmts}{numbers of compartment compartment associated with an observations, to derive concentration (typically obs_cmt)}
}
\value{
a matrix
}
\description{
Compute the derivatives
}
\examples{
derivatives(
v_DV = c(400, 40, 200, 20),
v_cmt = c(2, 3, 2, 3),
cmts = c(2,3)
)

}
