% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{est001}
\alias{est001}
\title{Estimation object 001}
\format{
An object of class \code{mapbayests} of length 10.
}
\source{
\url{https://github.com/FelicienLL/mapbayr-CPTPSP-2021}
}
\usage{
est001
}
\description{
Parameter estimation of 11 subjects from data001, done with model001.
}
\keyword{datasets}
