% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_data.mapbayests}
\alias{get_data.mapbayests}
\title{Return data from a mapbayests}
\usage{
\method{get_data}{mapbayests}(x, ...)
}
\arguments{
\item{x}{mapbayests object}

\item{...}{not used}
}
\value{
a tibble
}
\description{
Return data from a mapbayests
}
