% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_x}
\alias{get_x}
\alias{get_data}
\alias{get_eta}
\alias{get_cov}
\alias{get_param}
\alias{get_phi}
\title{Get content from object}
\usage{
get_data(x, ...)

get_eta(x, ...)

get_cov(x, ...)

get_param(x, ...)

get_phi(x, ...)
}
\arguments{
\item{x}{object to get content from}

\item{...}{passed along}
}
\value{
the class of the object returned depends on the function, and on their arguments. Typically, a data.frame or a vector if the output can be reduced to one line.
}
\description{
Helpful functions to get content from a \code{mrgmod} object (i.e. data) or from a \code{mapbayests} object (\code{data}, \code{eta}, \code{cov}, \code{param}, \code{phi}).
}
