% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve_model.R
\name{mbr_model_ref}
\alias{mbr_model_ref}
\title{Get model reference from mrgsolve model}
\usage{
mbr_model_ref(x)
}
\arguments{
\item{x}{model object}
}
\value{
model reference as a character string
}
\description{
Get model reference from mrgsolve model
}
\examples{
model <- mrgsolve::mread("ex_mbr1", mbrlib())
mbr_model_ref(model)

}
