% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploit_estimates.R
\name{mbrpred}
\alias{mbrpred}
\title{Compute full PK profile prediction from mapbayr estimates.}
\usage{
mbrpred(est, end = NULL, ...)
}
\arguments{
\item{est}{a list, default output of mbrest}

\item{end}{end of infusion time (passed to mrgsim)}

\item{...}{additional argument to pass to mrgsim}
}
\value{
a dataframe, a mrgsolve 'df' output
}
\description{
Compute full PK profile prediction from mapbayr estimates.
}
