% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelaveraging.R
\name{model_averaging}
\alias{model_averaging}
\title{Compute Weights from Several Estimations}
\usage{
model_averaging(..., scheme = c("LL", "AIC"), estlist = NULL)
}
\arguments{
\item{...}{estimation objects generated from \code{\link[=mapbayest]{mapbayest()}} to compute weight from}

\item{scheme}{scheme weight, either "LL" or "AIC"}

\item{estlist}{a list of estimation objects. Overrides \code{...}}
}
\value{
a matrix of numeric, the weight of the estimation. There is one row per subject ID, one column per model. Consequently, the sum of each row is 1. Named estimation objects (either in \code{...} or in \code{estlist}) will be used as column names in the output.
}
\description{
Model Averaging consists in analyzing the same data with different models
and to average their predictions.
In order to perform weigthed means of clearance predictions, (or
concentrations, or any metric of interest), it is necessary to compute
the "weight" of each estimation.
It is informed by the likelihood of estimation.
Two weighting scheme are currently implemented, one based on the log-
likelihood ("LL", the default), the other on the Akaike criterion ("AIC").
The method was previously described by Uster et al
\href{https://ascpt.onlinelibrary.wiley.com/doi/full/10.1002/cpt.2065}{(Clinical Pharmacology and Therapeutics, 2021)}.
}
\examples{
library(magrittr)
same_data_and_est <- function(x){
  x \%>\%
    adm_rows(ID = 2, time = 0, amt = 100, addl = 3, ii = 24) \%>\%
    obs_rows(ID = 2, time = 96, DV = 1) \%>\%
    adm_rows(ID = 9, time = 0, amt = 200, addl = 3, ii = 24) \%>\%
    obs_rows(ID = 9, time = 96, DV = 1) \%>\%
    mapbayest()
  }

mod <- exmodel(1, add_exdata = FALSE)

est1 <- mod \%>\%
  mrgsolve::param(TVCL = 2) \%>\%
  same_data_and_est()

est2 <- mod \%>\%
  mrgsolve::param(TVCL = 10) \%>\%
  same_data_and_est()

model_averaging(CL2 = est1, CL10 = est2)
model_averaging(estlist = list(A = est1, B = est2, est1))

}
