% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{obs_lines}
\alias{obs_lines}
\alias{obs_lines.data.frame}
\alias{obs_lines.missing}
\alias{obs_lines.mrgmod}
\title{Add observation lines to a dataset}
\usage{
obs_lines(x, ...)

\method{obs_lines}{data.frame}(x, ID = NULL, time, evid = 0L, cmt, DV = NA_real_, mdv = NULL, ...)

\method{obs_lines}{missing}(...)

\method{obs_lines}{mrgmod}(x, cmt = NULL, DV = NA_real_, DVmet = NULL, ...)
}
\arguments{
\item{x}{either a data.frame or a 'mrgsolve' model object}

\item{...}{additional columns}

\item{ID}{subject ID (default is 1)}

\item{time}{event time}

\item{evid}{event identification (default is 1 for administration, 0 for observation)}

\item{cmt}{compartment (no default, except if \verb{[OBS]} was tagged in the \verb{$CMT} block in model code. See \code{examples}.)}

\item{DV}{dependant value, i.e. observed concentration.}

\item{mdv}{missing dependant value (default is 0 a non-missing concentration value to take into account for parameter estimation, 1 otherwise)}

\item{DVmet}{second observation at the same time, often a metabolite ("DVmet") observed jointly with parent drug ("DV"). Works only if \code{x} is a 'mrgsolve' model where two \verb{[OBS]} compartments were defined (see \code{examples})}
}
\value{
a data.frame, or a 'mrgsolve' model with a dataset in the \verb{@args$data} slot (accessible with \code{get_data()}).
}
\description{
The \code{obs_lines()} function adds an one or several observation lines to a dataset provided as a proper data.frame or within a 'mrgsolve' model. Used in combination with \code{adm_lines()} and \code{add_covariates()}, it facilitates the creation of datasets in the proper format for simulations with 'mrgsolve' or parameter estimation with 'mapbayr'.
}
\examples{
# Create a dataset from scratch
obs_lines(time = 12, DV = 0.12, cmt = 2)

# Pipe-friendly addition of administration record to a pre-existing dataset
library(magrittr)
obs_lines(time = 12, DV = 0.12, cmt = 2) \%>\%
  obs_lines(time = c(24, 36, 48), DV = c(0.34, 0.56, 0.78), mdv = c(0,1,0), cmt = 2)

# Start from a 'mrgsolve' model
library(mrgsolve)
house() \%>\%
  obs_lines(time = 12, DV = 0.12, cmt = 2) \%>\%
  obs_lines(time = c(24, 36, 48), DV = c(0.34, 0.56, 0.78), mdv = c(0,1,0), cmt = 2) \%>\%
  mrgsim()

# Default observation compartments
# Set default observation compartments in the code with `[OBS]`
model <- mcode("model", "
$CMT @annotated
DEPOT : Depot
CENTR : Central [OBS]
", compile = FALSE)
obs_cmt(model)

# Thus, no need to manually specified `cmt = 2` anymore.
model \%>\%
  obs_lines(time = 12, DV = 0.12) \%>\%
  obs_lines(time = c(24, 36, 48), DV = c(0.34, 0.56, 0.78), mdv = c(0,1,0)) \%>\%
  get_data()

# Automatic lines duplication if parent + metabolite defined in the model
model <- mcode("model", "
$CMT @annotated
DEPOT : Depot
CENTR : Central [OBS]
PERIPH : Periph
METABO : Metabo [OBS]
", compile = FALSE)
obs_cmt(model)

model \%>\%
  obs_lines(time = 12, DV = 0.12, DVmet = 120) \%>\%
  obs_lines(time = c(24, 36, 48), DV = c(0.34, 0.56, 0.78), DVmet = c(340, 560, 780)) \%>\%
  get_data()

}
