% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{obs_lines}
\alias{obs_lines}
\title{Generate observation lines for a dataset}
\usage{
obs_lines(x, time, DV, mdv = 0, DVmet = NULL)
}
\arguments{
\item{x}{model object}

\item{time}{vector of time}

\item{DV}{vector of values to fit}

\item{mdv}{should the Dv be ignored (1) or not (0)}

\item{DVmet}{optional : metabolite data to fit}
}
\value{
model object with dataset
}
\description{
Generate observation lines for a dataset
}
