% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_post_process.R
\name{output_mbr}
\alias{output_mbr}
\title{Build the output of mbrest function}
\usage{
output_mbr(idata, model, arg.optim, arg.ofv, opt.value, post, output)
}
\arguments{
\item{idata}{data passed through processing}

\item{model}{a compiled mrgsolve_model}

\item{arg.optim}{argument passed to optimization function}

\item{arg.ofv}{argument passed to optimization function}

\item{opt.value}{value obtained by optimization function}

\item{post}{output of the post.process function}

\item{output}{return a mapbay_tab only}
}
\value{
a mbrests list object
}
\description{
Build the output of mbrest function
}
