% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbayests.R
\name{plot.mapbayests}
\alias{plot.mapbayests}
\title{Plot predictions from mapbayests object}
\usage{
\method{plot}{mapbayests}(x, ..., ci = FALSE)
}
\arguments{
\item{x}{A \code{mapbayests} object.}

\item{...}{additional arguments (not used)}

\item{ci}{print confidence interval (see \code{augment} documentation for more details)}
}
\value{
a \code{ggplot} object. Observed and predicted concentration vs time profile for every patients.
}
\description{
Plot predictions from mapbayests object
}
