% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbay_estimation.R
\name{postprocess}
\alias{postprocess}
\title{Post process results from optimization}
\usage{
postprocess(data, model, newuoa_value, data0, pre)
}
\arguments{
\item{data}{data passed through processing}

\item{model}{a compiled mrgsolve_model}

\item{newuoa_value}{output returned by newuoa (list of length 4)}

\item{data0}{original data}

\item{pre}{preprocessed object}
}
\value{
a list of post processing values
}
\description{
Post process results from optimization
}
