% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbrest_postprocess.R
\name{postprocess}
\alias{postprocess}
\alias{postprocess.optim}
\alias{postprocess.output}
\title{Postprocess mapbayr}
\usage{
postprocess.optim(x, data, opt.value)

postprocess.output(x, data, arg.optim, arg.ofv, opt.value, post, output)
}
\arguments{
\item{x}{the model object}

\item{data}{NMTRAN-like data set}

\item{opt.value}{value returned by optimizer}

\item{arg.optim, arg.ofv}{argument passed to optimizer}

\item{post}{output of the post.process function}

\item{output}{if \code{NULL} (the default) a mbrests object is returned; if \code{df} a \emph{mapbay_tab} dataframe is returned}
}
\description{
Functions to generate postprocess after optimization.
}
