% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbay_estimation.R
\name{postprocess}
\alias{postprocess}
\title{Post process results from optimization}
\usage{
postprocess(data, model, opt.value, arg.optim, arg.ofv)
}
\arguments{
\item{data}{data passed through processing}

\item{model}{a compiled mrgsolve_model}

\item{opt.value}{value obtained by optimization function}

\item{arg.optim}{argument passed to optimization function}

\item{arg.ofv}{argument passed to optimization function}
}
\value{
a list of post processing values
}
\description{
Post process results from optimization
}
