% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbayest_postprocess.R
\name{postprocess}
\alias{postprocess}
\alias{postprocess.optim}
\alias{postprocess.output}
\title{Postprocess mapbayr}
\usage{
postprocess.optim(x, data, opt.value, arg.ofv, arg.optim, hessian)

postprocess.output(
  x,
  arg.optim,
  arg.ofv.fix,
  arg.ofv.id,
  opt.value,
  post,
  output,
  times
)
}
\arguments{
\item{x}{the model object}

\item{data}{NMTRAN-like data set}

\item{opt.value}{value returned by optimizer}

\item{arg.optim, arg.ofv, arg.ofv.fix, arg.ofv.id}{argument passed to optimizer}

\item{hessian}{a function to compute the Hessian and variance-covariance matrix. Default is "optimHess", alternatively "nlmixrHess".}

\item{post, times}{output of the post.process function}

\item{output}{if \code{NULL} (the default) a mapbayests object is returned; if \code{df} a \emph{mapbay_tab} dataframe is returned}
}
\value{
\code{postprocess.optim()} returns a list with final parameters and \code{mapbay_tab}. \code{postprocess.output()} returns a \code{mapbayests} class object.
}
\description{
Functions to generate postprocess after optimization.
}
