% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbrest_check_and_preprocess.R
\name{preprocess.data}
\alias{preprocess.data}
\title{Preprocess: data into a list of individual data}
\usage{
preprocess.data(data)
}
\arguments{
\item{data}{NMTRAN-like data set}
}
\value{
a named list of data set (n individuals)
}
\description{
Preprocess: data into a list of individual data
}
