% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_post_process.R
\name{preprocess.ofv}
\alias{preprocess.ofv}
\title{Preprocess model and data for ofv computation}
\usage{
preprocess.ofv(model, data)
}
\arguments{
\item{model}{a compiled mrgsolve_model}

\item{data}{a dataframe, dataset (NM-TRAN format) of one individual to fit}
}
\value{
a list of argument passed to optimization function
}
\description{
Preprocess model and data for ofv computation
}
