% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_post_process.R
\name{preprocess.ofv}
\alias{preprocess.ofv}
\title{Preprocess: model and data for ofv computation}
\usage{
preprocess.ofv(x, data)
}
\arguments{
\item{x}{the model object}

\item{data}{NMTRAN-like data set}
}
\value{
a list of named arguments passed to optimizer (i.e. arg.ofv)
}
\description{
Preprocess: model and data for ofv computation
}
