% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbayests.R
\name{print.mapbayests}
\alias{print.mapbayests}
\title{Print a mapbayests object}
\usage{
\method{print}{mapbayests}(x, ...)
}
\arguments{
\item{x}{A \code{mapbayests} object.}

\item{...}{additional arguments}
}
\value{
print the results of the estimation to the console, and returns it invisibly.
}
\description{
Print a mapbayests object
}
