% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbayests.R
\name{use_posterior}
\alias{use_posterior}
\title{Use posterior estimation}
\usage{
use_posterior(x, covariance = FALSE, .zero_re = NULL)
}
\arguments{
\item{x}{A \code{mapbayests} object.}

\item{covariance}{Update the OMEGA matrix with the variance-covariance matrix of estimation (a logical, default is \code{FALSE}).}

\item{.zero_re}{Set all elements of the OMEGA or SIGMA matrix to zero. Default is "both" if \code{covariance} is FALSE, "sigma" otherwise. (possible values are "both", "sigma", "omega", "none")}
}
\value{
a mrgmod
}
\description{
Use posterior estimation
}
\details{
This function takes the results of an estimation (i.e. a \code{mapbayests} object) and return a modified \code{mrgmod} in order to perform \emph{a posteriori} simulations. Modifications are:
\itemize{
\item In $PARAM, the values of ETA are updated to the estimated values (instead of 0) and the covariates values are updated to the values of the individual (instead of default values)
\item If \code{covariance} is \code{TRUE}, the values of OMEGA are updated with the variance-covariance matrix of estimation (i.e. an approximation of the \emph{a posteriori} distribution) instead of the inter-individual variability (i.e. the \emph{a priori} distribution). Use this command in order to derive a confidence interval of concentrations that reflects the uncertainty about parameter estimation when a large number of profiles are simulated. Note that if inter-individual variability was initially defined in multiple $OMEGA blocks in the model, they will be collapsed to a single full matrix (this is irreversible).
\item Depending on the values of \code{.zero_re}, the elements of $OMEGA or $SIGMA can be set to zero, whether you want to simulate one profile, or several in order to derive confidence/prediction intervals.
This function works if there is only one individual in the dataset. It does not handle time-varying covariates.
}
}
